+++ r/R/methods_MASS.R
@@ -130,20 +130,7 @@ get_model_matrix.negbin <- function(model, newdata, mfx = NULL) {
 #' @noRd
 #' @export
 get_jacobian_analytic.negbin <- function(model, type, ...) {
-    if (
-        !identical(class(model)[1], "negbin") ||
-            !isTRUE(type %in% c("response", "link"))
-    ) {
-        return(NULL)
-    }
-    response_scale <- identical(type, "response")
-    jacobian_analytic_model_matrix(
-        model = model,
-        type = type,
-        response_scale = response_scale,
-        family = if (response_scale) stats::family(model) else NULL,
-        ...
-    )
+    jacobian_analytic_glm_family(model, "negbin", type, ...)
 }
 
 
@@ -157,16 +144,5 @@ get_model_matrix.rlm <- function(model, newdata, mfx = NULL) {
 #' @noRd
 #' @export
 get_jacobian_analytic.rlm <- function(model, type, ...) {
-    if (
-        !identical(class(model)[1], "rlm") ||
-            !identical(type, "response")
-    ) {
-        return(NULL)
-    }
-    jacobian_analytic_model_matrix(
-        model = model,
-        type = type,
-        response_scale = FALSE,
-        ...
-    )
+    jacobian_analytic_linear(model, "rlm", type, "response", ...)
 }