+++ r/R/methods_brglm2.R
@@ -40,18 +40,5 @@ get_model_matrix.brglmFit <- function(model, newdata, mfx = NULL) {
 #' @noRd
 #' @export
 get_jacobian_analytic.brglmFit <- function(model, type, ...) {
-    if (
-        !identical(class(model)[1], "brglmFit") ||
-            !isTRUE(type %in% c("response", "link"))
-    ) {
-        return(NULL)
-    }
-    response_scale <- identical(type, "response")
-    jacobian_analytic_model_matrix(
-        model = model,
-        type = type,
-        response_scale = response_scale,
-        family = if (response_scale) stats::family(model) else NULL,
-        ...
-    )
+    jacobian_analytic_glm_family(model, "brglmFit", type, ...)
 }